fix(capture): Win pointer_state 修正 GetAsyncKeyState/VK_LBUTTON 模块路径——在 Input::KeyboardAndMessaging 之外的 KeyboardAndMouse
CI windows job 暴露（本地 Linux 无法编译该路径）：GetAsyncKeyState 与
VK_LBUTTON 属 Win32::UI::Input::KeyboardAndMouse（SendInput 同源 feature
已开），且 VIRTUAL_KEY 是 u16 类型别名非 newtype。
本地已补 x86_64-pc-windows-msvc / aarch64-apple-darwin 交叉 check capture
（其余 crate 交叉受 openh264 C++ 工具链限制，交给 CI）

+++ crates/capture/src/other.rs
@@ -166,16 +166,20 @@ pub fn cursor_position() -> Option<(i32, i32)> {
 }
 
 /// Win：GetCursorPos + GetAsyncKeyState(VK_LBUTTON)，物理像素坐标系。
-/// GetAsyncKeyState 最高位 = 当前按住，无需窗口焦点。
+/// GetAsyncKeyState 最高位 = 当前按住，无需窗口焦点。注意 GetAsyncKeyState /
+/// VK_LBUTTON 在 KeyboardAndMessaging 之外的 Input::KeyboardAndMouse 模块
+/// （SendInput 同源，feature 已开）。
 #[cfg(windows)]
 pub fn pointer_state() -> Option<(i32, i32, bool)> {
     use windows_sys::Win32::Foundation::POINT;
-    use windows_sys::Win32::UI::WindowsAndMessaging::{GetAsyncKeyState, GetCursorPos, VK_LBUTTON};
+    use windows_sys::Win32::UI::Input::KeyboardAndMouse::{GetAsyncKeyState, VK_LBUTTON};
+    use windows_sys::Win32::UI::WindowsAndMessaging::GetCursorPos;
     let mut pt = POINT { x: 0, y: 0 };
     if unsafe { GetCursorPos(&mut pt) } == 0 {
         return None;
     }
-    let down = unsafe { GetAsyncKeyState(VK_LBUTTON) } as u16 & 0x8000 != 0;
+    // VK_LBUTTON 是 VIRTUAL_KEY（= u16 类型别名）；返回 i16 最高位 = 按下
+    let down = unsafe { GetAsyncKeyState(VK_LBUTTON as i32) } as u16 & 0x8000 != 0;
     Some((pt.x, pt.y, down))
 }
